requirements.txt
Flask
boto3
PyYAML
python-dotenv

